refactor(auth): move AuthController from Admin to root Controllers namespace
